frontend-utils: Move ExternalNavigator from Desktop
