+++ frontend-utils/src/backends.rs
@@ -1,2 +1,4 @@
 pub mod executor;
 pub mod storage;
+
+pub mod navigator;
